database matches fixing , handle disconnect game aborting , code optimisation
