+++ back/src/Game/Game.Module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { Game } from './Game';
 import { JwtModule } from '@nestjs/jwt';
+import { GameStartEvent } from './game-start-event';
 
 @Module({
   imports: [JwtModule.register({})],
-  providers: [Game],
+  providers: [Game, GameStartEvent],
 })
 export class GameModule {}
