+++ front/src/pages/game/Game.tsx
@@ -61,7 +61,6 @@ function useEffectOnce(effect: React.EffectCallback) {
 }
 
 const Game: React.FC = () => {
-    const token: string = localStorage.getItem('token') as string;
     const [user, setUser] = useState<UserType | null>(null);
     const [otherAvatar, setOtherAvatar] = useState<string | null>(null);
     const [otherUsername, setOtherUsername] = useState<string | null>(null);
@@ -73,7 +72,6 @@ const Game: React.FC = () => {
     const [ball, setBall] = useState<Ball | null>(null);
     const [otherpad, setOtherpad] = useState<Paddle | null>(null);
     const [socket, setSocket] = useState<MySocket | null>(null);
-    const [Dimensions, setDimention] = useState({ width: 0, height: 0 });
     const [init, setInit] = useState(false);
     const divRefs = {
         gameContainer: useRef<HTMLDivElement>(null),
@@ -89,7 +87,6 @@ const Game: React.FC = () => {
     const canvasRef = useRef<HTMLCanvasElement>(null);
     const ctx = canvasRef.current?.getContext('2d');
     const socketGame = React.useContext(SocketGameContext);
-    const renderData = React.useContext(RenderContext);
 
     useEffect(() => {
         const fetchUser = async () => {
@@ -127,10 +124,10 @@ const Game: React.FC = () => {
     }, 16.66);
 
     const setupEventListeners = () => {
-            canvasRef.current?.addEventListener(
-                'mousemove',
-                throttleHandleMouseMove
-            );
+        canvasRef.current?.addEventListener(
+            'mousemove',
+            throttleHandleMouseMove
+        );
     };
 
     useEffectOnce(() => {
@@ -189,9 +186,9 @@ const Game: React.FC = () => {
                 setOtherAvatar,
                 setGameOver,
                 setId,
-                setDimention,
                 setInit,
-                setOtherUsername
+                setOtherUsername,
+                setGameStarted
             );
             listning = true;
         }
@@ -212,41 +209,6 @@ const Game: React.FC = () => {
         socket?.emit('StartGameRobot');
         setGameStarted(true);
     };
-
-    // useEffect(() => {
-    //     if (padd && Dimensions.width > 0 && Dimensions.height > 0)
-    //         updateDivPosition(
-    //             divRefs.playerPaddle.current,
-    //             padd,
-    //             Dimensions.width,
-    //             Dimensions.height
-    //         );
-    //     console.log('padd', padd);
-    // }, [padd]);
-
-    // useEffect(() => {
-    //     if (padd && otherpad && canvasRef.current && ctx && ball && Dimensions.width > 0 && Dimensions.height > 0){
-
-    //         // updateDivPosition(
-    //             //     divRefs.ball.current,
-    //             //     ball,
-    //             //     Dimensions.width,
-    //             //     Dimensions.height
-    //             // );
-
-    //         }
-
-    // }, [ball]);
-
-    // useEffect(() => {
-    //     if (otherpad && Dimensions.width > 0 && Dimensions.height > 0)
-    //         updateDivPosition(
-    //             divRefs.otherPaddle.current,
-    //             otherpad,
-    //             Dimensions.width,
-    //             Dimensions.height
-    //         );
-    // }, [otherpad]);
 
     return (
         <div className="containerGame">
@@ -271,7 +233,7 @@ const Game: React.FC = () => {
                 </div>
             ) : null}
             <div className="container-profile">
-                {id === 1 ? (
+                {id !== 1 ? (
                     <>
                         <div className="player-profile">
                             <img
